test: split provider detection tests
